refactor/EV-804/componetização list coins
